requirements.txt
cryptography==3.4.8
pyjwt
requests
boto3
cachetools
